pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "inventory"
version = "0.1.0"
description = "Inventory service with product management API"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "Django>=5.0",
  "django-storages[boto3]>=1.14",
  "boto3>=1.34",
  "python-dotenv>=1.0",
  "psycopg[binary]>=3.2",
  "pillow>=10.0",
]

[project.urls]
Homepage = "https://example.com"

[tool.setuptools.packages.find]
include = ["inventory", "products"]
